requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
matplotlib>=3.4.0
pandas>=1.3.0
scikit-learn>=0.24.0
pillow>=8.2.0
pyyaml>=5.4.0
tqdm>=4.60.0
tensorboard>=2.5.0
pytest>=6.2.0
albumentations>=1.0.0  # For advanced augmentations
wandb>=0.12.0          # For experiment tracking (optional)